# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NGFW Management API
  version: 0.3.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /security/events:
    get:
      summary: Bounded M3 Suricata security-event stream
      parameters:
        - {name: session_id, in: query, schema: {type: string}}
        - {name: sensor_id, in: query, schema: {type: string, enum: [ids, ips]}}
        - {name: mode, in: query, schema: {type: string, enum: [IDS, IPS]}}
        - {name: severity, in: query, schema: {type: string, enum: [UNKNOWN, INFO, LOW, MEDIUM, HIGH, CRITICAL]}}
        - {name: verdict, in: query, schema: {type: string, enum: [UNKNOWN, ALERT, DROP, ERROR]}}
        - {name: correlation_state, in: query, schema: {type: string, enum: [CORRELATED, UNCORRELATED, AMBIGUOUS]}}
        - {name: application, in: query, schema: {type: string}}
        - {name: signature_id, in: query, schema: {type: integer, format: int64, minimum: 1, maximum: 4294967295}}
        - {name: after_sequence, in: query, schema: {type: integer, format: int64, minimum: 0}}
        - {name: stream_id, in: query, schema: {type: string, maxLength: 256}}
        - {name: cursor, in: query, description: Opaque next_cursor returned by the previous page; mutually exclusive with stream_id/after_sequence, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200':
          description: Bounded SecurityEventPage; gap/reset_required explicitly report evicted history or a new engine stream
          content: {application/json: {schema: {$ref: '#/components/schemas/SecurityEventPageEnvelope'}}}
        '400': {description: Invalid filter}
        '503': {description: Engine IPC or inspection runtime unavailable}
        '504': {description: Engine IPC timed out}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  schemas:
    SessionTuple:
      type: object
      required: [family, src_ip, dst_ip, protocol]
      properties:
        family: {type: string, enum: [ipv4, ipv6]}
        src_ip: {type: string}
        src_port: {type: integer, minimum: 0, maximum: 65535}
        dst_ip: {type: string}
        dst_port: {type: integer, minimum: 0, maximum: 65535}
        protocol: {type: integer, minimum: 1, maximum: 255}
        icmp_type: {type: integer, minimum: 0, maximum: 255}
        icmp_code: {type: integer, minimum: 0, maximum: 255}
        icmp_id: {type: integer, minimum: 0, maximum: 65535}
...
    ApplicationIdentity:
      type: object
      required: [name, source, confidence, conflicted]
      properties:
        name: {type: string, enum: [UNKNOWN, HTTP, TLS, DNS, SSH, OTHER]}
        raw_name: {type: string}
        source: {type: string, enum: [UNKNOWN, SURICATA_APP_PROTO, DPI_PARSER, TLS_METADATA, PORT_HEURISTIC]}
        confidence: {type: string, enum: [UNKNOWN, LOW, MEDIUM, HIGH, VERIFIED]}
        first_seen: {type: string, format: date-time}
        last_seen: {type: string, format: date-time}
        conflicted: {type: boolean}
    EnforcementResult:
      type: object
      properties:
        mechanism: {type: string, enum: [NONE, SURICATA_NFQUEUE, NFT_SESSION_GUARD]}
        scope: {type: string, enum: [PACKET, SESSION]}
        requested_action: {type: string}
        status: {type: string, enum: [NOT_REQUESTED, PENDING, REPORTED, APPLIED, FAILED, UNAVAILABLE]}
        reason: {type: string}
        observed_at: {type: string, format: date-time}
        operation_id: {type: string}
    ThreatEvent:
      type: object
      required: [event_id, sequence, ingested_at, event_class, source, capture_mode, correlation_state, correlation_revision, severity, application, verdict, enforcement]
      properties:
        event_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        observed_at: {type: string, format: date-time}
        ingested_at: {type: string, format: date-time}
        event_class: {type: string, enum: [security]}
        source: {type: string}
        sensor_id: {type: string, enum: [ids, ips]}
        sensor_epoch: {type: string}
        sensor_config_hash: {type: string}
        ruleset_id: {type: string}
        capture_mode: {type: string, enum: [IDS, IPS]}
        suricata_flow_id: {type: string, description: Kept as a decimal string so uint64 values remain exact in JavaScript}
        transaction_id: {type: string}
        observed_tuple: {$ref: '#/components/schemas/SessionTuple'}
        flow_tuple: {$ref: '#/components/schemas/SessionTuple'}
        protocol_raw: {type: string}
        session_id: {type: string}
        policy_id: {type: string}
        policy_generation: {type: integer, format: int64}
        correlation_state: {type: string, enum: [CORRELATED, UNCORRELATED, AMBIGUOUS]}
        correlation_reason: {type: string}
        correlation_revision: {type: integer, format: int64}
        signature_id: {type: integer, format: int64, minimum: 1, maximum: 4294967295}
        signature_revision: {type: integer, format: int64}
        signature: {type: string}
        category: {type: string}
        source_severity: {type: integer}
        severity: {type: string, enum: [UNKNOWN, INFO, LOW, MEDIUM, HIGH, CRITICAL]}
        application: {$ref: '#/components/schemas/ApplicationIdentity'}
        signature_action: {type: string}
        packet_verdict: {type: string}
        verdict: {type: string, enum: [UNKNOWN, ALERT, DROP, ERROR]}
        enforcement: {$ref: '#/components/schemas/EnforcementResult'}
    SecurityEventPageEnvelope:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean}
        data:
          type: object
          required: [items, stream_id, next_sequence, next_cursor, has_more, oldest_sequence, gap, evicted_count]
          properties:
            items: {type: array, maxItems: 200, items: {$ref: '#/components/schemas/ThreatEvent'}}
            stream_id: {type: string}
            gap_from: {type: integer, format: int64}
            next_sequence: {type: integer, format: int64}
            next_cursor: {type: string}
            has_more: {type: boolean}
            oldest_sequence: {type: integer, format: int64}
            gap: {type: boolean}
            reset_required: {type: boolean}
            evicted_count: {type: integer, format: int64}
...